requirements.txt
Django==3.0.5
django-mongoengine==0.4.4
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.12.2
djangorestframework-csv==2.1.0
djangorestframework-xml==2.0.0
djongo==1.3.3
mongoengine==0.20.0
pymongo==3.11.1
urllib3==1.26.2
Unidecode==1.1.1
drf_yasg==1.20.0